Following changes: - added updateBaseProof(), tokenAmountCalc(), getBlockRewardFactor() - L1_NO_GAS() error - rewardIssued to uint256 - eip1559.py + several debug prints

+++ packages/protocol/contracts/L1/TaikoData.sol
@@ -14,7 +14,7 @@ library TaikoData {
 
     struct FeeAndRewardConfig {
         uint256 baseFeeProof;
-        uint64 rewardIssued;
+        uint256 rewardIssued;
         // avgProofTimeMAF is known already
         //uint16 avgProofTimeMAF;
         uint64 rewardTargetPerGas;
